Rename swarm<->horde so code matches meaning
swarm = the single entity standing in for N zombies (was 'horde unit'/cluster);
horde = the telegraphed attack wave (was 'swarm event'). Full rename, save-break
approved:
- lib/horde.lua (cluster) <-> lib/swarm.lua (wave) swapped; tests likewise.
- tiers.HORDE* -> tiers.SWARM*; entity prototypes zomtorio-horde-* -> zomtorio-swarm-*.
- storage keys, module identifiers, comments updated to match.
- settings: zomtorio-horde-size-multiplier -> zomtorio-zombie-count-multiplier;
  zomtorio-swarm-events/intensity/frequency -> zomtorio-horde-events/intensity/frequency
  (config fns + locale display names updated).
- Kept (already correct): zomtorio-swarm-melee tech/damage type, zomtorio-nest-swarm-*,
  /zomtorio-horde command.
- CLAUDE.md updated to final naming.
