+++ locale/en/zomtorio.cfg
@@ -5,9 +5,9 @@ zomtorio-corpse-kiln=Corpse kiln
 
 [entity-name]
 zomtorio-corpse-kiln=Corpse kiln
-zomtorio-horde-small=Zombie swarm
-zomtorio-horde-medium=Zombie swarm
-zomtorio-horde-big=Zombie swarm
+zomtorio-swarm-small=Zombie swarm
+zomtorio-swarm-medium=Zombie swarm
+zomtorio-swarm-big=Zombie swarm
 
 [recipe-name]
 zomtorio-kiln-dried-corpse=Kiln-dried corpse
@@ -36,7 +36,7 @@ zomtorio-reanimation-minutes=Corpse reanimation time (minutes)
 zomtorio-bot-collect-corpses=Bots collect corpses
 zomtorio-pollution-cost-multiplier=Pollution cost per zombie (x vanilla)
 zomtorio-nest-spawn-rate=Nest spawn rate
-zomtorio-horde-size-multiplier=Horde size multiplier
+zomtorio-zombie-count-multiplier=Zombie count multiplier
 zomtorio-zombie-cap=Active zombie cap
 zomtorio-nest-swarm-base=Nest swarm size (low pollution)
 zomtorio-nest-swarm-max=Nest swarm size (max pollution)
@@ -47,9 +47,9 @@ zomtorio-player-infection=Player infection
 zomtorio-night-speedup=Night speedup (fraction added)
 zomtorio-expansion-rate=Base expansion rate
 zomtorio-night-assault-multiplier=Night assault escalation
-zomtorio-swarm-events=Escalating swarm events
-zomtorio-swarm-intensity=Swarm event intensity
-zomtorio-swarm-frequency=Swarm event frequency
+zomtorio-horde-events=Escalating horde events
+zomtorio-horde-intensity=Horde intensity
+zomtorio-horde-frequency=Horde frequency
 
 [mod-setting-description]
 zomtorio-corpse-reanimation=Killed zombies drop corpses that reanimate into new zombies after a timer (via the spoilage system).
@@ -57,7 +57,7 @@ zomtorio-reanimation-minutes=How long a corpse takes to reanimate. Bakes into th
 zomtorio-bot-collect-corpses=Mark dropped corpses for deconstruction so construction/logistic bots haul them to storage. They still reanimate there unless burned or kiln-dried first.
 zomtorio-pollution-cost-multiplier=Lower means each unit of pollution recruits far more (individually weak) attackers. Bakes into prototypes.
 zomtorio-nest-spawn-rate=How much faster nests produce enemies than vanilla. Bakes into prototypes.
-zomtorio-horde-size-multiplier=Overall multiplier on the number of zombies generated.
+zomtorio-zombie-count-multiplier=Overall multiplier on the number of zombies generated.
 zomtorio-zombie-cap=Maximum number of individual (real) active zombies for CPU tuning. Overflow folds into abstract horde clusters rather than being lost.
 zomtorio-nest-swarm-base=When the active zombie cap is full, a nest's extra output folds into a local swarm cluster. This is how big that swarm grows at a nest with little pollution.
 zomtorio-nest-swarm-max=The largest a single nest's local swarm grows when the nest sits in heavy pollution. Stops an un-triggered nest from growing an endless swarm.
@@ -68,6 +68,6 @@ zomtorio-player-infection=Being bitten infects the player; any net healing clear
 zomtorio-night-speedup=How much faster zombies move at night. 1.0 means +100% (double daytime speed). Startup setting: changing it takes effect after a restart.
 zomtorio-expansion-rate=How aggressively enemy bases expand and spread.
 zomtorio-night-assault-multiplier=How much larger/more frequent assaults are at night.
-zomtorio-swarm-events=Telegraphed massive assaults at night whose size and frequency grow with evolution.
-zomtorio-swarm-intensity=Scales the spawn rate and duration of swarm events.
-zomtorio-swarm-frequency=Scales how often swarm events occur.
+zomtorio-horde-events=Telegraphed massive attack waves (hordes) at night whose size and frequency grow with evolution.
+zomtorio-horde-intensity=Scales the spawn rate and duration of horde attack waves.
+zomtorio-horde-frequency=Scales how often hordes attack.
