+++ test/zomtorio-tests/tests/nest_spec.lua
@@ -12,7 +12,7 @@
 
 local T     = require("harness.runner")
 local nest  = require("__zomtorio__.lib.nest")
-local horde = require("__zomtorio__.lib.horde")
+local swarm = require("__zomtorio__.lib.swarm")
 local tiers = require("__zomtorio__.lib.tiers")
 
 --- Spawn a real enemy biter near `pos` to stand in for one a spawner just emitted.
@@ -25,14 +25,14 @@ end
 --- The (single) tracked cluster near a position.
 local function find_cluster(surface, pos)
   local found = surface.find_entities_filtered {
-    name = tiers.HORDE.small, position = pos, radius = 32,
+    name = tiers.SWARM.small, position = pos, radius = 32,
   }
   return found[1]
 end
 
 local function reset(cap)
-  horde.set_cap_override(cap)
-  horde.reset_state()
+  swarm.set_cap_override(cap)
+  swarm.reset_state()
   nest.set_budget_override(nil)
 end
 
@@ -53,10 +53,10 @@ T.test("under the cap, nest output stays a tracked individual", function(t)
   local biter = make_biter(t.surface, o)
   t.assert.not_nil(biter, "test biter created")
 
-  local before = horde.active_count()
+  local before = swarm.active_count()
   nest.on_entity_spawned { entity = biter }
   t.assert.is_true(biter.valid, "the individual is kept (not folded) under the cap")
-  t.assert.equal(before + 1, horde.active_count(), "it now counts against the cap")
+  t.assert.equal(before + 1, swarm.active_count(), "it now counts against the cap")
 end)
 
 -- Cap full: nest output is destroyed and folded into a LOCAL cluster at the nest,
@@ -72,11 +72,11 @@ T.test("over the cap, nest output folds into a growing local cluster", function(
   t.assert.is_false(b1.valid, "the over-cap individual is removed")
   local cluster = find_cluster(t.surface, o)
   t.assert.not_nil(cluster, "a local cluster formed at the nest")
-  t.assert.equal(1, horde.pop_of(cluster), "cluster holds the folded zombie")
+  t.assert.equal(1, swarm.pop_of(cluster), "cluster holds the folded zombie")
 
   local b2 = make_biter(t.surface, o)
   nest.on_entity_spawned { entity = b2 }
-  t.assert.equal(2, horde.pop_of(cluster), "the next spawn grows the same local cluster")
+  t.assert.equal(2, swarm.pop_of(cluster), "the next spawn grows the same local cluster")
 end)
 
 -- Cap full AND the local swarm at its budget: the nest is saturated, so further
@@ -93,13 +93,13 @@ T.test("at the nest budget, further output is throttled (no infinite swarm)", fu
   local b2 = make_biter(t.surface, o); nest.on_entity_spawned { entity = b2 }
   local cluster = find_cluster(t.surface, o)
   t.assert.not_nil(cluster, "cluster formed")
-  t.assert.equal(2, horde.pop_of(cluster), "cluster filled to the budget")
+  t.assert.equal(2, swarm.pop_of(cluster), "cluster filled to the budget")
 
   -- One more: local swarm (2) >= budget (2) -> throttled, cluster unchanged.
   local b3 = make_biter(t.surface, o)
   nest.on_entity_spawned { entity = b3 }
   t.assert.is_false(b3.valid, "the saturating spawn is dropped")
-  t.assert.equal(2, horde.pop_of(cluster), "the swarm does not grow past its budget")
+  t.assert.equal(2, swarm.pop_of(cluster), "the swarm does not grow past its budget")
 end)
 
 -- Non-enemy or off-Nauvis spawns are ignored (R-SCOPE-1 / friendly units untouched).
@@ -112,8 +112,8 @@ T.test("a non-enemy unit is ignored", function(t)
     name = tiers.INDIVIDUAL.small, position = o, force = "player",
   }
   t.assert.not_nil(biter, "player-force unit created")
-  local before = horde.active_count()
+  local before = swarm.active_count()
   nest.on_entity_spawned { entity = biter }
   t.assert.is_true(biter.valid, "left untouched")
-  t.assert.equal(before, horde.active_count(), "not counted against the cap")
+  t.assert.equal(before, swarm.active_count(), "not counted against the cap")
 end)
